glispのソース構成を大幅に見直した
