+++ glisp/src/fractal/koch.rs
@@ -5,23 +5,22 @@
    hidekuno@gmail.com
 */
 use crate::draw::DrawLine;
+use crate::fractal::Fractal;
 
 pub struct Koch {
     sin60: f64,
     cos60: f64,
-    scale: i64,
     draw_line: DrawLine,
 }
 impl Koch {
-    pub fn new(c: i64, draw_line: DrawLine) -> Koch {
+    pub fn new(draw_line: DrawLine) -> Koch {
         Koch {
             sin60: ((std::f64::consts::PI * 60.0) / 180.0).sin(),
             cos60: ((std::f64::consts::PI * 60.0) / 180.0).cos(),
-            scale: c,
             draw_line: draw_line,
         }
     }
-    pub fn draw(&self, x0: f64, y0: f64, x1: f64, y1: f64, c: i64) {
+    pub fn draw(&self, x0: f64, y0: f64, x1: f64, y1: f64, c: i32) {
         if c > 1 {
             let xa = (x0 * 2.0 + x1) / 3.0;
             let ya = (y0 * 2.0 + y1) / 3.0;
@@ -39,27 +38,32 @@ impl Koch {
             (self.draw_line)(x0, y0, x1, y1);
         }
     }
-    pub fn do_demo(&self) {
+}
+impl Fractal for Koch {
+    fn get_func_name(&self) -> &'static str {
+        "draw-koch"
+    }
+    fn do_demo(&self, c: i32) {
         self.draw(
             0.3597222222222222,
             0.0,
             0.04722222222222222,
             0.6964285714285714,
-            self.scale,
+            c,
         );
         self.draw(
             0.04722222222222222,
             0.6964285714285714,
             0.6708333333333333,
             0.6964285714285714,
-            self.scale,
+            c,
         );
         self.draw(
             0.6708333333333333,
             0.6964285714285714,
             0.3597222222222222,
             0.0,
-            self.scale,
+            c,
         );
     }
 }
