+++ glisp/src/fractal/mod.rs
@@ -8,3 +8,8 @@ pub mod dragon;
 pub mod koch;
 pub mod sierpinski;
 pub mod tree;
+
+pub trait Fractal {
+    fn get_func_name(&self) -> &'static str;
+    fn do_demo(&self, c: i32);
+}
